+++ opensearch-20171225/src/Models/ReleaseSortScriptResponseBody/result.php
@@ -1,0 +1,48 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\OpenSearch\V20171225\Models\ReleaseSortScriptResponseBody;
+
+use AlibabaCloud\Dara\Model;
+
+class result extends Model
+{
+    /**
+     * @var int
+     */
+    public $version;
+    protected $_name = [
+        'version' => 'version',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->version) {
+            $res['version'] = $this->version;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['version'])) {
+            $model->version = $map['version'];
+        }
+
+        return $model;
+    }
+}