+++ opensearch-20171225/src/Models/UpdateSortScriptRequest.php
@@ -6,32 +6,14 @@ namespace AlibabaCloud\SDK\OpenSearch\V20171225\Models;
 
 use AlibabaCloud\Dara\Model;
 
-class CreateSortScriptRequest extends Model
+class UpdateSortScriptRequest extends Model
 {
     /**
      * @var string
      */
     public $description;
-
-    /**
-     * @var string
-     */
-    public $scope;
-
-    /**
-     * @var string
-     */
-    public $scriptName;
-
-    /**
-     * @var string
-     */
-    public $type;
     protected $_name = [
         'description' => 'description',
-        'scope' => 'scope',
-        'scriptName' => 'scriptName',
-        'type' => 'type',
     ];
 
     public function validate()
@@ -44,18 +26,6 @@ class CreateSortScriptRequest extends Model
         $res = [];
         if (null !== $this->description) {
             $res['description'] = $this->description;
-        }
-
-        if (null !== $this->scope) {
-            $res['scope'] = $this->scope;
-        }
-
-        if (null !== $this->scriptName) {
-            $res['scriptName'] = $this->scriptName;
-        }
-
-        if (null !== $this->type) {
-            $res['type'] = $this->type;
         }
 
         return $res;
@@ -71,18 +41,6 @@ class CreateSortScriptRequest extends Model
         $model = new self();
         if (isset($map['description'])) {
             $model->description = $map['description'];
-        }
-
-        if (isset($map['scope'])) {
-            $model->scope = $map['scope'];
-        }
-
-        if (isset($map['scriptName'])) {
-            $model->scriptName = $map['scriptName'];
-        }
-
-        if (isset($map['type'])) {
-            $model->type = $map['type'];
         }
 
         return $model;