+++ opensearch-20171225/ChangeLog.md
@@ -1,3 +1,11 @@
+2025-12-29 Version: 4.0.0
+- Update API CreateSortScript: add request parameters body.description.
+- Update API GetSortScript: add response parameters Body.result.description.
+- Update API ReleaseSortScript: add response parameters Body.result.
+- Update API UpdateSortScript: add request parameters body.
+- Update API UpdateSortScript: add request The number of query or body parameters has changed from zero to many.
+
+
 2025-12-22 Version: 3.0.2
 - Update API ModifyAppGroupQuota: add request parameters clientToken.